Changed nfq handling
Got rid of class NFQChannel, since using select is just as simple
and more readable

Wrapped nfq.unbind with a try catch block. Unbinding the queue sometimes
fails. Unbinding successfuly is however not required to close the queue

Fixed a segfault due to an attempt to set a verdict using an
already closed NF queue