Modified breakpoint / pausing mechanism, add command next
Continue now simply merge the primary and temporary lists,
without trying to trigger breakpoints on each packets received
during pause.

The user cannot see the content (or existence) of the temporary list.

The first breakpoint of a run session will trigger automatically.
The user will have to call the command next to trigger a (manual) breakpoint
on the following packets